More tech stuff done...
git-svn-id: svn://192.168.1.10/public/CoffeeMud@14111 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Libraries/CMGenEditor.java
@@ -8501,7 +8501,7 @@ public class CMGenEditor extends StdLibrary implements GenericEditor
 			{
 				final TechComponent E=(TechComponent)me;
 				E.setInstalledFactor((float)prompt(mob, E.getInstalledFactor(), ++showNumber, showFlag, L("Installed Factor")));
-				E.setRechargeRate(prompt(mob, E.getRechargeRate(), ++showNumber, showFlag, L("Pwr Recharge Rate")));
+				E.setRechargeRate((float)prompt(mob, E.getRechargeRate(), ++showNumber, showFlag, L("Pwr Recharge Rate")));
 			}
 			if(me instanceof ShipEngine)
 			{
@@ -8912,7 +8912,7 @@ public class CMGenEditor extends StdLibrary implements GenericEditor
 			{
 				final TechComponent E=(TechComponent)me;
 				E.setInstalledFactor((float)prompt(mob, E.getInstalledFactor(), ++showNumber, showFlag, "Installed Factor"));
-				E.setRechargeRate(prompt(mob, E.getRechargeRate(), ++showNumber, showFlag, L("Pwr Recharge Rate")));
+				E.setRechargeRate((float)prompt(mob, E.getRechargeRate(), ++showNumber, showFlag, L("Pwr Recharge Rate")));
 			}
 			if(!(me instanceof Electronics))
 			{
